move settings to local files
