+++ tests/performance/locust_files/metamodeling/workflow.py
@@ -163,3 +163,12 @@ class MetaModelingUser(HttpUser):
             file_uuid = response.json().get("id")
         assert file_uuid is not None
         return UUID(file_uuid)
+
+
+if __name__ == "__main__":
+    from locust_settings import LocustSettings, dump_dotenv
+
+    class MetaModelingSettings(UserSettings, LocustSettings):
+        pass
+
+    dump_dotenv(MetaModelingSettings())
