--- tests/performance/locust_files/platform_ping_test/_settings.py
@@ -1,10 +1,0 @@
-from locust_settings import LocustSettings, dump_dotenv
-from platform_ping_test import LocustAuth
-
-
-class LoadTestSettings(LocustAuth, LocustSettings):
-    pass
-
-
-if __name__ == "__main__":
-    dump_dotenv(LoadTestSettings())
