+++ tests/performance/locust_files/platform_ping_test/platform_ping_test.py
@@ -47,3 +47,12 @@ class WebApiUser(FastHttpUser):
 
     def on_stop(self):  # pylint: disable=no-self-use
         print("Stopping locust user")
+
+
+if __name__ == "__main__":
+    from locust_settings import LocustSettings, dump_dotenv
+
+    class LoadTestSettings(LocustAuth, LocustSettings):
+        pass
+
+    dump_dotenv(LoadTestSettings())
